Get all implemented for shops

+++ src/main/java/com/sellio/model/result/PageData.java
@@ -1,9 +1,6 @@
 package com.sellio.model.result;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.util.List;
 
@@ -13,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 public class PageData<T> {
     private int totalPages;
-    private int totalElements;
+    private long totalElements;
     private boolean firstPage;
     private boolean lastPage;
     private int size;
